Have proper init data and ddl-auto update.
Since no method is left to initialize data, create would destroy the
data in dev environment without an easy way to put it back in DB.

The next development steps should (for a time) not try to modify
anything about the DB, so it's not a big issue atm. It will be in the
future, see #13, #14, and even #11 to some extent.